% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArea.R
\name{getArea}
\alias{getArea}
\title{getArea}
\usage{
getArea(object, group, type = "all")
}
\arguments{
\item{object}{Either an object of class NPSForVeg, or a list of objects of class NPSForVeg}

\item{group}{Indicates which plant groups the area should correspond to, as different groups are monitored on different size plots. The name of the group should be enclosed in quotes. Acceptable values are" "trees", "saplings", "seedlings","shrubs", "shseedlings","vines", "herbs".}

\item{type}{Indicates the type of area to be returned, and should be enclosed in quotes. Acceptable values are
\describe{
 \item{"all"}{(the default) Returns the total area sampled per plot for that group.}
 \item{"single"}{Returns the area of one of the subsamples (eg. microplot or quadrat).If there is only one sample of that group, such as trees sample over an entire plot, than "single" and "all" will produce identical results}
 \item{"count"}{A count of the number of the number of subplots/microplots/quadrats that the group is monitored in.}
}}
}
\value{
For a single NPSForVeg object getArea returns a vector of length one. For a list of such object it returns a vector as long as the list.
}
\description{
Retrieves the number and area of various plot sizes from an object of class NPSForVeg.
}
\examples{
\dontrun{
 netn <- importNETN("C:/NETN/R_Dev/data/NPSForVeg/NETN")

# single microplot area
sap_area <- getArea(netn, group = "saplings", type = "single")

# full sapling area sampled
sap_area <- getArea(netn, group = "saplings", type = "all")

# Num. sapling plots
sap_count <- getArea(netn, group = "saplings", type = 'count')

# area for seedlings
seed_area <- getArea(netn, group = "seedlings", type = 'all')

# Tree plot area per park
park_plot_area <- cbind(area = getArea(netn, group = "trees"),
                        parks = getNames(netn, name.class = 'code'))
}

}
