% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCycles.R
\name{getCycles}
\alias{getCycles}
\title{getCycles}
\usage{
getCycles(object, cycles = NA, years = NA)
}
\arguments{
\item{object}{Either a data.frame like that stored in the \code{Cycles} slot, an \code{NPSForVeg} object or a list of such objects.}

\item{cycles}{A vector of cycle codes. Will only return data from those cycles.}

\item{years}{A length 1 numeric vector containing one or more years. Will only return cycles that include those years.}
}
\value{
Either a \code{data.frame} or a \code{list} containing the contents of the "Cycles" slot for one or more \code{NPSForVeg} objects.
}
\description{
Gets the cycles \code{data.frame} from one or more \code{NPSForVeg} objects.
}
\details{
When the "object" is a \code{data.frame} or a single \code{NPSForVeg} object this function will return the \code{data.frame} in 
the "Cycles" \code{slot}. This \code{data.frame} will be filtered as specified in the other arguments. When "object" is a \code{list} 
then a \code{list} is returned. Each element of the \code{list} will be the filtered \code{data.frame} from "Cycles" \code{slot} from
the input \code{NPForVeg} objects.
}
