% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getNames.R
\name{getNames}
\alias{getNames}
\title{getNames}
\usage{
getNames(object, name.class = "short")
}
\arguments{
\item{object}{Either an NPSForVeg object or a list of such objects}

\item{name.class}{Tyep of name to return. One of three options, in quotes.
\describe{
\item{"short"}{ The default. Returnes the short name of the park}
\item{"long"}{Returns the long, format name of the park}
\item{"code"}{Returns the park code}
}}
}
\value{
A character vector with one or more park names.
}
\description{
Retreives park names from an NPSForVeg object of a list of such objects.
}

