% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getNetwork}
\alias{getNetwork}
\title{getNetowrk}
\usage{
getNetwork(object)
}
\arguments{
\item{object}{Either a NPSForVeg object, or a list of such objects}
}
\value{
A characeter vector of network codes. For a single NPSForVeg object a single code is returned. For a list of objects, a network code for each object in the list is returned.
}
\description{
Returns the network code for a object of class NPSForVeg
}

