% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSoilChemNames.R
\name{getSoilChemNames}
\alias{getSoilChemNames}
\title{getSoilChemNames}
\usage{
getSoilChemNames(object, chem.names = "chemvar")
}
\arguments{
\item{object}{Either an NPSForVeg object or a list of such objects}

\item{chem.names}{Type of name to return. One of two options, in quotes.
\describe{
\item{"chemvar"}{ The default. Returnes the short name of the soil chemistry variable}
\item{"varlong"}{Returns the long, format name of the soil chemistry variable}
}}
}
\value{
A character vector with one or more soil chemistry names.
}
\description{
Retreives soil chemsitry variable names from an NPSForVeg object of a list of such objects.
}

