% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSoils.R
\name{getSoils}
\alias{getSoils}
\title{getSoils}
\usage{
getSoils(object, values, cycles = NA, years = NA, plots = NA,
  output = "dataframe")
}
\arguments{
\item{object}{Either an object of class \code{NPSForVeg} or a list of such objects}

\item{values}{Defaults to \code{NA}. A required character string indicating which horizon should be selected. Options are "Ohor" or "Ahor".}

\item{cycles}{Defaults to \code{NA}. A numeric vector indicating which cycles should be included. This is determined by matching the appropriate \code{Cycle} field to \code{cycles}}

\item{years}{Defaults to \code{NA}. A numeric vector indicating which years should be included. This is determined by matching the appropriate \code{Sample_Year} field to \code{years}}

\item{plots}{Defaults to \code{NA} A character vector indicating which plots should be included. This is determined by macthing the appropriate \code{Plot_Name} field to \code{plots}.}

\item{output}{Either "dataframe" or "list". Determines the output type When \code{object} is a list. "Dataframe", the default, indicates the output from all of \code{NSPForVeg} objects should be combined into a single \code{data.frame}. "List" will return a \code{list} where each element of the list is a \code{data.frame} from a single \code{NPSForVeg} object, and each element is named based on that objects \code{ParkCode} slot.}
}
\description{
Retrieves soil chemistry data from an \code{NPSForVeg} object.
}
\details{
This function extracts data on soil chemistry from an \code{NPSForVeg} object. This function is called by other analysis and graphing functions.
}

