% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubplotCount.R
\name{getSubplotCount}
\alias{getSubplotCount}
\title{getSubplotCount}
\usage{
getSubplotCount(object, group, years = NA, plots = NA, subtype = "all",
  ...)
}
\arguments{
\item{object}{Either an NPSForVeg object of a list of such objects.}

\item{group}{Indicates which plant groups the area should correspond to. Acceptable values are "saplings","seedlings", and "herbs"}

\item{subtype}{Indicates the type of area to be returned, and should be enclosed in quotes. Acceptable values are 
\describe{
 \item{"all"}{(the default) Returns the total area and number of subplots sampled per plot event for that group.}
 \item{"area"}{(the default) Returns the total area sampled per plot event for that group.}
 \item{"count"}{A count of the number of subplots/microplots/quadrats that the group is monitored in for each plot event.}
}}
}
\value{
For a single NPSForVeg object getSubplotCount returns a data.frame.
}
\description{
This function gets the the counts of quadrats or microplots sampled in the "Events" slot of an NPSForVeg object.
}
