% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importERMN.R
\name{importERMN}
\alias{importERMN}
\title{importERMN}
\usage{
importERMN(Dir)
}
\arguments{
\item{Dir}{The directory where the data is found}
}
\value{
Returns a list containing 8 \code{NPSForVeg} objects, one for each park, named using the standard 4 letter park code (e.g. ALPO, BLUE etc.) as part of a list.
}
\description{
This function imports data from the standard ERMN .csv files and saves it as \code{NPSForVeg} objects. The required .csv files are: Plots, Events, Trees, Saplings, Seedlings, ERMNSoil,SoilChemistryVariables,and CommonNames.
}

