% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMIDN.R
\name{importMIDN}
\alias{importMIDN}
\title{importMIDN}
\usage{
importMIDN(Dir)
}
\arguments{
\item{Dir}{The directory where the data is found}
}
\value{
Returns a list with 12 \code{NPSForVeg} objects, one for each park.
}
\description{
This function imports data from the standard MIDN .csv files and saves it as \code{NPSForVeg} objects. The required .csv files are: Plots, Events, Trees, Saplings, Seedlings, Vines and CommonNames.
}

