% Generated by roxygen2 (4.0.1): do not edit by hand
\name{importNCRN}
\alias{importNCRN}
\title{importNCRN}
\usage{
importNCRN(Dir)
}
\arguments{
\item{Dir}{The directory where the data is found}
}
\value{
Returns 11 NPSForveg objects, one for each park, named using the standard 4 letter park code (e.g. ANTI, CATO etc.) and a list, named NCRN, which contians all 11 objects.
}
\description{
This function imports data from the standard NCRN .csv files and saves it as NPSOForVeg objects. The required .csv files are: Plots, Events, Trees, Saplings, Seedlings, Shrubs, Shrub_Seedlings, Vines, Herbs and CommonNames.
}

