% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSHEN.R
\name{importSHEN}
\alias{importSHEN}
\title{importSHEN}
\usage{
importSHEN(Dir)
}
\arguments{
\item{Dir}{The directory where the data is found. Path should not have a trailing slash.}
}
\value{
Returns 1 \code{NPSForVeg} object for the park.
}
\description{
This function imports data from Shenandoah's .csv files and saves it as an \code{NPSForVeg} object. The required .csv files are: Plots, Events, Trees, Saplings, Seedlings, Shrubs, Shrub_Seedlings, Herbs and CommonNames.
}
