% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVplot.R
\name{IVplot}
\alias{IVplot}
\title{IVplot}
\usage{
IVplot(
  object,
  top = NA,
  IVargs = NA,
  parts = FALSE,
  compare = NA,
  labels = NA,
  colors = c("darkgreen", "green", "yellow"),
  ...
)
}
\arguments{
\item{object}{Either a \code{data.frame} which is the output of \code{\link{IV}}, or an object of class \code{NPSForVeg}, or a \code{list} of objects of class \code{NPSForVeg}.}

\item{top}{Defaults to \code{NA}. An integer. Indicates the number of species that should be included from each dataset in the graph (see \code{compare}). For a single dataset the species with the largest total IV values are selected. When multiple datasets are compared, the top species in each dataset are selected, which can result in more species than \code{top} included in the final graph - e.g. if \code{top=3} and two datasets are used,  up to six species can be included in the graph, depeding on the amount of overlap it the top three species. If \code{top=NA} all species are displayed.}

\item{IVargs}{A required \code{list}, defaults to \code{NA}. The arguments in this list are passed to \code{\link{IV}} and are used to generate the data for the graph. Any argument that is valid for \code{IV} can be in this list, and should be written in arguement=value format e.g. \code{IVargs=list(group="trees", years=c(2010:2013))}}

\item{parts}{Logical value, defaults to \code{FALSE}. Indicates if the components of IV (Density, Size and Distribution) should be shown on the graph.}

\item{compare}{A \code{list} of lists. Defaults to \code{NA}. This can be used to compare the IV values from several different datasets, such as trees from different parks, different time periods,  or trees, sapings and seedlings from the same park. Each list will be passed on to \code{\link{IV}}, and can contain any arguments that are valid for that function. Entries in the list should be written in argument=value format  e.g. \code{compare=list( list(object=CATO, group="saplings", years=c(2010:2013)), list(object=CATO, group="seedlings", years=c(2010:2013)) )}. Any number of datasets can be added using \code{compare}.}

\item{labels}{A character vector, defaults to \code{NA}. Labels for each dataset, only used when \code{compare} is not \code{NA}}

\item{colors}{Colors for the graph.}

\item{...}{Other arguments passed to \code{barchart} in the \pkg{lattice} package.}
}
\value{
Returns a barchart
}
\description{
Makes a barchart of IV values
}
\details{
This function uses the \code{barchart} function in the \pkg{lattice} package to create a horizontal barchart that displays IV values. The format of the display varies depending on the type of data to be displayed. If only a single dataset is considered (\code{compare=NA}), and \code{parts=FALSE}, then a simple horizontal bar chart is produced. If \code{parts=TRUE} a stacked bar chart is produced. Each component of IV: density, size and distribution, is shown in a different color.

If two or more datasets are used, and \code{parts=FALSE} then all results are graphed in a single panel. Each dataset is a different color, and the legend uses the contents of \code{labels} to identify each dataset. If \code{parts=TRUE} then each dataset is plotted in a separate panel, and the panel title is determined by \code{labels}. Within each panel, the three components of IV will each have a separate color, which will be consistent between panels.
}
\examples{
\dontrun{
midn <- importMIDN("C:/Data/")
# APCO tree IV plot
IVplot(midn[[1]], top = 10, IVargs = list(group = "trees", years = 2021:2024))

COLO seedling IV plot
IVplot(midn[[4]], top = 10, IVargs = list(group = 'seedlings', years = 2022:2024))

}


}
\seealso{
\code{\link{IV}}
}
