% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapPlants.R
\name{mapPlants}
\alias{mapPlants}
\title{mapPlants}
\usage{
mapPlants(object, plots, values, maptype = "basic", colorgroups = 8,
  radius = 30, opacity = 1, colortype = "quantile", colors = c("cyan",
  "magenta4", "orangered3"), ...)
}
\arguments{
\item{object}{Either an object of class \code{NPSForVeg} or a list of such objects}

\item{plots}{A character vector of plot names.}

\item{values}{A vector of values to map. Can be either numberical data or categories.}

\item{maptype}{The type of base map to be used. Options are:
 \describe{
 \item{"basic"}{The default, uses the basic Park Tiles map, similar to the maps found in park brochures.}
 \item{"imagery"}{The Park Tiles satellite imagery map.}
 \item{"slate"}{The Park Tiles slate map. A very muted and grey base map.}
}}

\item{colorgroups}{The number of different colors to display for maps with a \code{colortype} of "bin" or "quantile", or a series of cut points to be used with \code{colortype="bin"} See discussion below.}

\item{radius}{The radius in meters, of the circle drawn on the map. When the map is zoomed out the circles will be visible regardless of the value chosen.}

\item{opacity}{Opacity of the circles on the map. A single number from 0 (completely transparent) to 1 (completely opaque). Defaults to 1.}

\item{colortype}{Indicates the method of assiging colors to each plot. Make use of the leaflet package's color funcitons:
\describe{
  \item{"quantile"}{Divides the data into the number of groups indicated in \code{colorgroups} based on quantiles of the data. An approximately equal number of plots will be in each group.  Each group will be given a different color.}
  \item{"bin"}{Divides the data into groups based on the \code{cut} function. Like the "quantile" option in divides the plots into groups, but in this case each group should cover an approximately equal range in values of the data. Alternatively if \code{colorgroups} is a vector rather than a single number, the elements of the vector will be the cut points separating the groups to be mapped.}
  \item{"numeric"}{Colors the plots based on a smooth color ramp, rather than dividing into groups.}
  \item{"factor}{Used when \code{values} are categorical data.}
}}

\item{colors}{A character vector of one or more colors for the plots on the map. See discussion below.}
}
\description{
Produces an html map from a vector of plot names and a vector of corresponding values.
}
\details{
This function is serves as a wrapper for the leaflet package. It quickly creates a map by plotting the locations of the plots on the ParkTiles base map from the NPS.

Several option exist for coloring the plots based on the \code{colorNumeric}, \code{colorBin}, \code{colorQuantile}, and \code{colorFactor} functions in leaflet.

quantile - This option will divide the plots into the equal quantiles based on the data in \code{values}. The \code{colorgroups} agrument indicates the number of gourps the plots will be divided into. For exmple if \code{colorgroups=4} then the plots with the bottom quarter of \code{values} will get one color, the second quarter of \code{values} will get a second color etc. An erorr may occur if it is not possible to divide the \code{value} vector in to the indicate number of groups. For example, if half of \code{values} is the same number, than it cannot be divided into thirds.

bin - This option also divides the plots into groups, but in this case the groups either cover an equal range in \code{values} or cover ranges specificed in \code{colorgroups}. If \code{colorgroups} is a single value, than that will indicate the number of groups to divide \code{values} into. For example, if \code{values} range from 1 to 100, and \code{colorgroups} is 4, the plots with values 1-25 will be one color, 26-50 a second and so on. If instead \code{colorgroups} is \code{c(-1,10,50,75,101)} then plots from 1-10 will be one color, 11-50 a second and so on.

numeric - This option does not divide the plots into groups. Rather plots will have a smooth ramp of colors from the lowest to the highest value.

factor - This well color the plots when \code{values} is categorical data (e.g. soil or vegetation type) rather than numeric. Each category gets a different color.

The \code{colors} agrument indicates which colors will be chosen for the plots. This is a charcter vector and can either be standard R color names ("blue", "green" etc.) or hexadecimal colors ("#0000FF","008000",etc). If the number of colors is equal to the number of groups that each group will get the corresponding color. If there are fewer colors than groups, than a colorramp will be created using the indicate colors and used for the plots. Typically maps made with the \code{"numeric"} option will make use of the color ramp.

Currently there is no legend displayed with the map. However, if you click on a circle it will give you the Plot Name and the value associated with the plot.
}

