% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPSForVeg_Class_def.R
\docType{class}
\name{NPSForVeg-class}
\alias{NPSForVeg-class}
\title{S4 Class Definition for NPSForveg}
\description{
|An S4 class that contains the data from forest monitoring from a set of forest montoirng plots. This will typcially be a single park, but it could be any group of plots, including part of a park or a group of parks.
}
\section{Slots}{

\describe{
\item{\code{ParkCode}}{A short code to designate the park (or group of plots), typically an NPS 4 letter code. Stored as a length 1 character vector.}

\item{\code{ShortName}}{A short name for the park. Stored as a length 1 character vector.}

\item{\code{LongName}}{A long, formal name for the park. Stored as a length 1 character vector.}

\item{\code{Network}}{The code for the I&M network the park belongs to. Stored as a length 1 character vector.}

\item{\code{TPlotSize}}{A length 2 numeric vector. The first element is the number of subplots where trees are monitored at each plot. The second element is the size of each of the subplots in meters squared.}

\item{\code{SapPlotSize}}{A length 2 numeric vector. The first element is the number of microplots where saplings are monitored at each plot. The second element is the size of each of the microplots in meters squared.}

\item{\code{SeedPlotSize}}{A length 2 numeric vector. The first element is the number of quadrats where seedlings are monitored at each plot. The second element is the size of each of the quadrats in meters squared.}

\item{\code{ShrubPlotSize}}{length 2 numeric vector. The first element is the number of microplots where shrbus are monitored at each plot. The second element is the size of each of the microplots in meters squared.}

\item{\code{ShSeedPlotSize}}{A length 2 numeric vector. The first element is the number of quadrats where shrub seedlings are monitored at each plot. The second element is the size of each of the quadrats in meters squared.}

\item{\code{VPlotSize}}{A length 2 numeric vector. The first element is the number of subplots where vines on trees are monitored at each plot. The second element is the size of each of the subplots in meters squared.}

\item{\code{HPlotSize}}{A length 2 numeric vector. The first element is the number of quadrats where herbaceous plants are monitored at each plot. The second element is the size of each of the quadrats in meters squared. Note - these need not be actually herbaceous plants, but it is assumed that cover of each species is measured rather than the dbh or height.}

\item{\code{CWDTransectSize}}{a data.frame with information (including percent slope and slope-length) on the CWD transects}

\item{\code{Plots}}{a data.frame with information on the plots}

\item{\code{Events}}{a data.frame with information on each sampling event}

\item{\code{Trees}}{a data.frame with  tree data}

\item{\code{Saplings}}{a data.frame with sapling data}

\item{\code{Seedlings}}{a data.frame with seedling data}

\item{\code{Shrubs}}{a data.frame with shrub data}

\item{\code{ShSeedlings}}{a data.frame with shrub seedling data}

\item{\code{Vines}}{a data.frame with vine data}

\item{\code{Herbs}}{a data.frame with herbaceous plant data}

\item{\code{CWD}}{a data.frame with coarse woody debris data}

\item{\code{Commons}}{a data.frame which links common names to Latin names}
}}

