% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcIV}
\alias{calcIV}
\title{calcIV}
\usage{
calcIV(InCount, InSize, InPres)
}
\arguments{
\item{InCount}{A \code{data.frame} which is a site x species matrix where each cell is the number of individuals of a given species in a given site}

\item{InSize}{A \code{data.frame} like InCount except each cell is a measure of the size of the speices such as total basal area or total seedling height}

\item{InPres}{a \code{data.frame} which is a presence absence matrix}
}
\value{
Returns a \code{data.frame} with the Importance Values for each species, the columns are "Species" with the name of each species, "Density" which is density IV, "Size" which is size IV, "Distribution" which is distribution IV and "Total" which is \eqn{Density + Size +Distribution}
}
\description{
| Calculates the Importance Value from three Site X Species data frames
| One for abundance, one for basal area, one for presence-absence
}

