% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCommons}
\alias{getCommons}
\title{getCommons}
\usage{
getCommons(object, output = "dataframe")
}
\arguments{
\item{object}{Either a NPSForVeg object or a list of such objects.}

\item{output}{Used when "object" is a list to determine if the output is a data.frame or a list. Can be either "dataframe" (the default), or "list". Needs to be in quotes.}
}
\value{
Either a data.frame or a list containing the contents of the "Commons" slot for one or more NPSForVeg objects.
}
\description{
gets common names data.frame from one or more NPSForVeg objects.
}
\details{
When the "object" is a single NPSForVeg object this funciton will return the data.frame in the "Commons" slot  which should have the Latin and common names of the plants. When "object" is a list and "output" is "dataframe" a combied data.frame with the "Commons" slot from all NPSForVeg elements in returned. If the "object" is a list and "output" is "list" then a list is returned. Each element of the list will be the "Commons" slot from one the input NPForVeg objects, and the name of each element corresponds to the name of the NPSForVeg object it comes from.
}

