% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetwork.R
\name{getNetwork}
\alias{getNetwork}
\title{getNetwork}
\usage{
getNetwork(object)
}
\arguments{
\item{object}{Either a \code{NPSForVeg} object, or a \code{list} of such objects}
}
\value{
A character vector of network codes. For a single \code{NPSForVeg} object a single code is returned. For a \code{list} of objects, a network code for each object in the \code{list} is returned.
}
\description{
Returns the network code for a object of class \code{NPSForVeg}
}
\examples{
\dontrun{
netn <- importNETN("C:/NETN/R_Dev/data/NPSForVeg/NETN")

nwk_pk <- cbind(network = getNetwork(netn),
                parks = getNames(netn, name.class = 'code'))

}

}
