% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlantNames.R
\name{getPlantNames}
\alias{getPlantNames}
\title{getPlantNames}
\usage{
getPlantNames(object, names, out.style = "common", in.style = "Latin",
  output = "vector")
}
\arguments{
\item{object}{An \code{NPSForVeg} object}

\item{names}{a character vector of Latin names, common names or TSN numbers (as a character vector) to be translated}

\item{out.style}{Either "common", "Latin", or "TSN". Indicates the type of name that should be returned. Defaults to "common"}

\item{in.style}{Either "common", "Latin", or "TSN". Indicates the type of name found in the "names" argument.}

\item{output}{Currently does nothtin Will accepts "vector" or "list". Determines the output type When \code{object} is a list 
When implemented "list" will return a \code{list} where each element of the list is a \code{data.frame} from a
 single \code{NPSForVeg} object, and each element is named based on that object's \code{ParkCode} slot.}
}
\value{
A character vector of names in the format specified by \code{out.style}.
}
\description{
Translates between Latin names, common names and TSN numbers
}
\details{
This function will convert the names in \code{names} to the style specified in \code{out.style},
 using the data in the \code{Commons} slot of an object of class \code{NPSForVeg}. Note that if the data in the \code{Commons} slot
 has more than one entry for one or more of the input names, the fucniton will fail with an error.
}
