% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getPlantNames}
\alias{getPlantNames}
\title{getPlantName}
\usage{
getPlantNames(object, names, style = "common")
}
\arguments{
\item{object}{an \code{NPSForVeg} object}

\item{names}{a character vector of Latin names to be translated}

\item{style}{Currently only accepts "common", which is the default. May eventually allow a translation from common to Latin names}
}
\value{
A character vector of common names
}
\description{
Translates Latin names to common names
}
\details{
This funciton will match the Latin names in \code{names} with common names, using the data in the \code{Commons} slot of an object of class \code{NPSForVeg}.
}

