% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPlantNames.R
\name{getPlantNames}
\alias{getPlantNames}
\title{getPlantNames}
\usage{
getPlantNames(object, names, out.style = "common", in.style = "Latin",
  output = "dataframe")
}
\arguments{
\item{object}{An \code{NPSForVeg} object}

\item{names}{a character vector of Latin Names, common names or TSN numbers (as a character vector) to be translated}

\item{out.style}{Either "common", "Latin", or "TSN". Indicates the type of name that should be returned. Defaults to "common"}

\item{in.style}{Either "common", "Latin", or "TSN". Indicates the type of name found in the "names" argument.}

\item{output}{Currently only accpets"dataframe". Determines the output type When \code{object} is a list. "dataframe", the default, indicates the   output from all of \code{NSPForVeg} objects should be combined into a single \code{data.frame}. When implemented "list" will return a   \code{list} where each element of the list is a \code{data.frame} from a single \code{NPSForVeg} object, and each element is named based    on that objects \code{ParkCode} slot.}
}
\value{
A character vector of names in the format speceified by \code{out.style}.
}
\description{
Translates Latin names to common names using TSN numbers
}
\details{
This function will covert the names in \code{names} to the style specified in \code{out.style}, using the data in the \code{Commons} slot of an object of class \code{NPSForVeg}. Note that TSN numbers are used to covert from one style of name to the other.
}

