% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPlotNames.R
\name{getPlotNames}
\alias{getPlotNames}
\title{getPlotNames}
\usage{
getPlotNames(object, ..., type = "active", visits = NA, years = NA,
  subparks = NA, output = "vector")
}
\arguments{
\item{object}{Either an NPSForVeg object or a list of such objects}

\item{type}{One of three options that indicate what type of plots are to be considered. Must be in quotes. Options are:
\describe{
\item{"active"}{The default. Only returns names of plots which are listed as active in the \code{Plots$Location_Status} field.}
\item{"all"}{The default. Returns names from all types of plots.}
\item{"retired}{Only returns names from plots which are listed as retired in the \code{Plots$Location_Status} field.}
}}

\item{visits}{A numeric vector. Returns only plot names where the number of plot visits matches one of the values in \code{visits} The number of visits to a plot is determined by the \code{Event_Count} column in the \code{Events} slot.}

\item{years}{A numeric vector. Returns only plot names where the yeare the plot was visited  matches one of the values in \code{years} The year a visit takes place is determined by the \code{Event_Year} column in the \code{Events} slot.}

\item{subparks}{A character vector. Returns only plot names where the subpark the plot is in matches one of the values in \code{subparks} The subpark a plot is locaed in is determined by the \code{SubUnit_Code} column in the \code{Plots} slot.}

\item{output}{Either "vector" (the default) or "list". Note that this must be in qutoes. Determines the type of output from the function}
}
\description{
Returns the names of plots from an NPSForVeg object that meet various criteria.
}
\details{
This function returns plot names either from a single NPSForVeg object or a list of such objects. One use for this function is to assemble a vector of plot names that meet some criteria (location, when they were monitored, etc) that can then be passed on to the \code{plots} arguement of a second fucntion such as \code{getPlants}, \code{IV}, \code{dens} etc.  The default output is a charcacter vector. However, if \code{object} is a list and \code{output} is "list" then a list of character vectors will be returned. The name of each element in this list will correspond to the \code{ParkCode} in each NPSForVeg object.
}

