% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlots.R
\name{getPlots}
\alias{getPlots}
\title{getPlots}
\usage{
getPlots(object, type = "active", visits = NA, years = NA, plots = NA,
  subparks = NA, output = "dataframe")
}
\arguments{
\item{object}{An NPSForVeg object or a list of such objects.}

\item{type}{One of three options that indicate what type of plots are to be considered. Must be in quotes. Options are:
\describe{
\item{"active"}{The default. Only returns data for plots which are listed as active in the /code{Plots$Location_Status} field.} 
\item{"all"}{The default. retruns data from all types of plots.}
\item{"retired"}{Only returns data from plots which are listed as retired in the \code{Plots$Location_Status} field. }
}}

\item{visits}{A numeric vector. Returns only data from plots where the number of plot visits matches one of the values in \code{visits} The number of visits to a plot is determined by the \code{Event_Count} column in the \code{Events} slot.}

\item{years}{A numeric vector. Returns only plot data from plots where the yeare the plot was visited  matches one of the values in \code{years} The year a visit takes place is determined by the \code{Event_Year} column in the \code{Events} slot.}

\item{subparks}{A character vector. Returns only data from plots where the subpark the plot is in matches one of the values in \code{subparks} The subpark a plot is locaed in is determined by the \code{SubUnit_Code} column in the \code{Plots} slot.}

\item{output}{Either "dataframe" (the default) or "list". Note that this must be in quotes. Determines the type of output from the function.}
}
\description{
Returns the contents of the \code{Plots} slot of an NPSForVeg object. The retunred data can be filtered to meet various criteria.
}
\details{
This function returns plot data either from a single NPSForVeg object or a list of such objects. The default output is a data.frame. However, if \code{object} is a list and \code{output} is "list" then a list of data.frames will be returned. The name of each element in this list will correspond to the \code{ParkCode} in each NPSForVeg object.
}

