% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make}
\alias{make}
\title{make}
\usage{
make(object, ParkCode, ShortName, LongName, Network, plots = NA, ...)
}
\arguments{
\item{object}{Either an NPSForVeg object or a list of such objects}

\item{ParkCode}{The parkcode for the new NPSForVeg object}

\item{ShortName}{The short name for the new NPSForVeg object}

\item{LongName}{The long name for the new NPSForVeg object}

\item{Network}{The network code for hte new NPSForVeg object}

\item{plots}{A charcter vector with plot names. When specificied, only data from the given plots will be included in the new NPSForVeg Object.}
}
\description{
makes a new NPSForVeg object from one or more existing objects
}
\details{
This function creates new \code{NPSForVeg} objects by combining two or more previously existing objects and/or by subsetting exisitng objects. If more than one object is provided then the data for these objects is combined. The areas associated with the plots (eg. the size and number of sapling microplots) is taken from the first object. WARNING:  If the objects have different sized plots then area based calcuations will be in error. Providing a \code{plots} argument will indicate which plots are in the new object. The user must indicate the new network code, park code and park names.
}

