% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteXSpec.R
\name{SiteXSpec}
\alias{SiteXSpec}
\title{SiteXSpec}
\usage{
SiteXSpec(
  object,
  group,
  years = NA,
  cycles = NA,
  values = "count",
  status = "alive",
  area = "plot",
  output = "dataframe",
  species = NA,
  plots = NA,
  plot.type = "active",
  Total = TRUE,
  common = F,
  ...
)
}
\arguments{
\item{object}{either an object of class \code{NPSForVeg} or a list of such objects}

\item{group}{A required character string indicating which group of plants should be selected. Options are: "trees", "saplings",
"seedlings", "shrubs", "shseedlings" (indicating shrub seedlings), "vines", "herbs" or "cwd".}

\item{years}{Defaults to \code{NA}. A numeric vector indicating which years should be included. This is passed on to \code{\link{getPlants}}.}

\item{cycles}{Defaults to \code{NA}. A numeric vector indicating which cycles should be included. This is passed on to \code{\link{getPlants}}.}

\item{values}{Determines the data contained in the Site X Species matrix. Possible values are:
\describe{
\item{"count"}{The default. Each cell will include a count of the number of a given plant species in a given plot. For trees, saplings, seedlings, shrubs and shrub seedlings this is the number of plants. For vines, it is the number of trees a vine species grows on. For herbs it will be the number of quadrats the plant occurs in.}
\item{"size"}{For trees and saplings this is the total basal area in m2 per ha. For tree seedlings and shrub seedlings it is the total height, and for herbs it is the average percent cover across all sampled quadrats. For shrubs and vines there is no defined size and the function will terminate with an error.}
\item{"presab"}{Produces a presence/absence matrix. When a plant species is present in a given plot the corresponding cell value will 1, otherwise it is 0.}
}}

\item{status}{A required character string indicating if user wants data from living or dead plants. Used only for trees, saplings and shrubs. Values of this argument are matched to the \code{Status} field in the \code{Tree}, \code{Saplings} or \code{Shrubs} slot.  Acceptable options are:
\describe{
\item{"alive"}{The default. Includes any plant with a status of "Alive", "Alive Standing", "Alive Broken", "Alive Leaning" ,"Alive Fallen","AB","AF","AL","AM","AS","RB","RF","RL","RS" or "TR"}
\item{"snag"}{Standing dead tree only, Includes any plant with a status of "Dead", "Dead - Human Action", "Dead Leaning", "Dead Missing", "Dead Standing", "Dead - Too Small","DB","DL","DM",or"DS" }
\item{"other"}{Includes any plant with a status of "Missing", "Missing - Presumed Dead", "Missing - Uncertain" , "Downgraded to Non-Sampled","ES","EX","NL","XO", "XP" or"XS" }
\item{"all"}{Includes all plants}
}}

\item{area}{A character vector. Indicates if the values in the output are on a per plot or per area basis. This only works with counts,
basal areas, and length of seedlings.
\describe{
\item{"plot"}{The default. Each cell will be on a per plot basis. The only option for \code{values="presab"}.}
\item{"ha"}{Values will be on a per hectare basis}
\item{"ac"}{values will be on a per acre basis}
}}

\item{output}{Either "dataframe" or "list". Determines the output type when \code{object} is a list. "dataframe", the default, indicates the output from all of \code{NSPForVeg} objects should be a single large \code{data.frame}, containing all sites and species from all \code{NPSForVeg} objects. "list" will return a \code{list} where each element of the list is a \code{data.frame} from a single \code{NPSForVeg} object, and each element is named based on that object's \code{ParkCode} slot.}

\item{species}{A character vector of names. Defaults to \code{NA}. When not \code{NA} only species included in \code{species} will be included in the matrix. If a name is in \code{species}, but is not present in the data, it will not appear in the output.}

\item{plots}{A character vector of plot names. Defaults to \code{NA}. When not \code{NA} only plots included in \code{plots} will be included in the matrix. If a plot name is in \code{plots}, but is not present in the data, it will not appear in the output.}

\item{plot.type}{Passed on to \code{\link{getPlotNames}} and used to filter events so that only events from certain plot types are returned.
One of three options are availbe (always ineclosed in quotes).
\describe{
\item{"all"}{Events from all types of plots are returned.}
\item{"active"}{The default. Only returns events from plots which are listed as active in the Plots$Location_Status field.}
\item{"retired}{Only returns events from plots which are listed as retired in the Plots$Location_Status field. }
}}

\item{Total}{Logical value. Determine if a "Total" column will be included in the output. Defaults to TRUE.}

\item{common}{Defaults to \code{FALSE}. Indicates if common names should be used rather than Latin names.}

\item{...}{Other arguments passed on to \code{\link{getPlants}}}
}
\description{
Produces a Site X Species matrix. Each cell can contain a count, a measure of size or 1/0 indicating presence/absence
}
\details{
This function will first call \code{\link{getPlants}} to retrieve the plant data and \code{\link{getEvents}} to retrieve the event data. Then a site X species matrix will be created.
Values in the cells are determined by the option selected with \code{values}. Each row corresponds to a different plot and each column
to a different species.

Note that \code{species} can be a vector of common or Latin names. If common names are used then \code{common=TRUE} must be included in
the function call.
}
\examples{
\dontrun{
ncrn <- importNCRN("C:/Data")
trees4yr <- SiteXSpec(ncrn, group = "trees", years = 2021:2024)
saps4yr <- SiteXSpec(ncrn, group = "saplings", years = 2021:2024)

midn <- importMIDN("C:/Data/")
vafo <- midn[[13]]
vafo_herb4 <- SiteXSpec(vafo, group = "herbs", years = 2022:2024)

}

}
